ocr_model/requirements.txt
kserve==0.11.0
pillow==10.0.1
pytesseract==0.3.10